requirements.txt
streamlit
gspread
pandas
google-auth
google-auth-oauthlib
google-api-python-client
openpyxl
